pyproject.toml
[tool.poetry]
name = "linguadoc"
version = "0.1.0"
description = "DOC generator for language learning"
authors = ["Phoenix Grey <phoenix.grey0108@gmail.com>"]
repository = "https://github.com/qishe-nlp/linguadoc"
readme = "README.md"
documentation = "https://qishe-nlp.github.io/linguadoc/"
keywords = ["doc", "language", "auto"] 

[tool.poetry.dependencies]
python = "^3.8"
python-docx = "^0.8.11"
googletrans = "4.0.0rc1"
click = "^8.0.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
Sphinx = "^4.2.0"
sphinx-rtd-theme = "^1.0.0"

[tool.poetry.scripts]
preprocess_data = 'linguadoc._entry:preprocess_data'

gen_lingua_docx = 'linguadoc._entry:gen_lingua_docx'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
